+++ Changelog.md
@@ -61,6 +61,7 @@
 - Head damage is increased. ✔️
 - Perfect Aim mercenary skill is added. ✔️
 - Bounty system with rewards is added. ✔️
+- Bounty command is added now players are able to set new bounties. ✔️ 
 - Profession Commands Reworked. ✔️
 - Account Commands Rework. 20% 
 - Player Commands Rework. 0%